Use a bundler.
This means that packages such as microtime that involve native code
cannot be included here: the enclosing LF project that uses reactor-ts
will instead have to declare those dependencies.

This change was motivated by the huge (hundreds of MB) amount of disk
space required even for the most trivial LF programs. Nearly all of
that cruft is totally unused by LF programs and by the LF runtime,
and the bundler takes advantage of that fact by including only the
code that we need. After bundling, the runtime is a few orders of
magnitude smaller.