Attempt to fix failure in SlowingClockPhysical.
nextEvent holds the head of the event queue from BEFORE the reaction was
executed. Now that the reaction was executed, it might have enqueued
more recent events. In this case, shutdown had been the next event on
the event queue when in fact, the next event should have been an event
on the physical action a. If I'm right about this, it's pretty strange
that this error was not caught already.