Enqueue network output control reactions into `_reactionQ`
The level of each reaction should be checked before do that reaction.
However, it cannot be achieved if we execute output control reactions at
the end of the logical tag. Thus, I made the reactionQ as 'protected' to
push output control reactions into it. It is safe because only the
App and FederatedApp can access that Q.